dev-requirements.txt
pytest-ckan
djlint
isort
black
